DEV-905 Check for shallow dependencies rather than just cleaning up
This change will remove the indiscriminate cleanup of shallow runs.
Instead  we filter the dependent runs on their ini, if this is a
shallow run and it has other shallow dependencies we don't cleanup.

This allows support of the case where multiple shallow runs share the
same sample (ie one blood to multiple shallow tumors).