DEV-947 Eliminate hardcoded input/output strings
Many of our tests were hardcoding the strings for the input and
output directories which frustrates maintenance and is error-prone.

References to these strings have been centralised into the
`CommonTestEntities` class and some convenience methods added. This
approach is dead-simple (so as to not create test code that's too
smart and which runs the risk of re-implementing the implementation)
but still effective at the goal of removing duplication and hard-
coding.

Those substages which invoke Tabix have been reparented under a
common ancestor class which eliminates their identical re-
implementations of a test method for testing Tabix invocation.

Also an abstraction class, `BucketInputOutput` has been added
to simplify assertion against input and output locations that
are pushed or pulled by `gsutil`.