DEV-949 Delete instances with local SSDs upon non-success
The request to stop the VMs when they don't succeed is silently
getting dropped because VMs with local SSDs cannot actually be
stopped, because they cannot be started a second time.

Updated the code to assume that any instance started with local SSDs
should be deleted upon completion, regardless of outcome.