DEV-943 Create a common interface for stages of pipeline
Ticket gives the rationale. This commit creates an interface for all
stages and a common test hierarchy.

The interface contains a bunch of template methods, for the various
things you need to do to define a stage. Then the StageRunner ties all
this together.

One thing that is pretty awkward still is when the output file is
dynamically generated from a substage pipeline. This occurs in Somatic
and Germline caller, and needs some thought.

Instead of requiring to wire stages into fields, you can now simply
instantiate them in a pipeline's run method and pass them to the stage
runner. This makes the wiring in PipelineMain simpler.

Each stage should have a test which extends StageTest. It's also
template style and should guide the developer into testing all parts of
the stage. Currently the duplication in tools directories is removed
using an indirection to a string variable. Versions are duplicated.

These tests could potentially replace substage and command tests, as
they cover all bash statements.