DEV-861 Add bachelor to the pipeline
Adds a new stage bachelor, run at the same time as linx and health
checker. Bachelor takes purple, the tumor bams and germline caller as
input.

As germline caller runs in the single sample pipeline, a new "storage"
class is necessary to wait if necessary for the germline output. The new
class consolidates duplication with bam metrics, but not with alignment
yet (as it does not yet conform to the "stage" interface).

Bachelor will not be run in the smoke test, as germline calling still
takes too long to run in travis. This will be improved as part of a
separate task