DEV-890 Workaround for empty VCFs
When the RepeatMasker script is invoked against a valid VCF that
nonetheless consists of only metadata and the header line (i.e.
no data lines) it errors out. It is probably never invoked against
such a VCF in production but for Pv5 that situation occurs in
the smoke tests.

For now work around this situation by not calling the script at
all if the VCF contains no data lines.

Also timestamps have been added to the run logs in the lines that
indicate what is about to be run.