DEV-976 Remove explicit setting of temp dir
There was a bug in the structural caller that was happening
after the switch to /data/tmp for the temporary directory.

The GRIDSS commands were receiving an explicit TMP_DIR parameter
which was set to the wrong value, and which was causing the
(small) root filesystem on local-SSD-backed instances to fill
up when processing some samples.

Eliminate the explicit setting of the GRIDSS parameter and
instead export an environment variable at the top of the BASH
startup script to set the default for all JVM instances to
the desired directory.