DEV-976 Amend Rscript path to match image
In the development environment the path to the GRIDSS scripts has been
changed and that was not yet reflected in this version of the code
because a couple of key changes have been cherry-picked to fix a
production bug.

Also the Dockerfile has been updated as the Debian project has updated
its mirrors to not include openjdk-8-jre, and that has finally trickled
down to our image provider (so now we'll use version 11 instead).