DEV-965 Add BAM assertions to smoke test
Repurposes the BAM modules assertion framework for the smoke test. The
smoke test will now download BAMs and ensure they have the same
alignments, duplicates and sorting as an expected BAM.

Duplicated the classes so not to run into dependency hell creating new
dependencies between bam and cluster (hadoop mainly). The plan will be
to delete the bam module soon.