DEV-966 Add MANIFEST for output files
A manifest file is now written before pushing the files from a completed run
into the final storage location. This file contains a checksum, size in bytes
and full path (relative to the bucket) of the file.

The smoke test now checks the contents of the output directory against a
capture of the manifest as it currently looks to avoid regressions, and also
checks the contents of the manifest in the source cloud storage against the
contents of the remote cloud storage.

Suspect we may want to move some of the functionality that is currently in the
smoke test into the implementation. There should be some good discussion on
that in the PR.