DEV-968 Validate the fastq filename before parsing lanes
We have a dependency on the file naming of fastq files which is a bit
hidden. If a file was to not follow the convention you'd get odd
failures (like index out of bounds).

This change adds explicit validation of the file naming and fails hard
on a clear exception if the file is unparseable.