DEV-890 Resolve merge error
There was an erroneous merge after the original completion of
DEV-890 that was only caught during manual validation. In the
meantime many things changed with the structure of the code
so the merge essentially had to be re-done from scratch.

Also a bug was found in the wrapper script for the repeat
masker which was fixed separately.