DEV-889 Eliminate full pass in preprocess stage
Change to computing the insertsizemetrics on a subset of the input BAMs
per advice from Daniel. The thinking is that this will be good enough,
and will save a full pass over each of the input BAMs.

Also remove some of the old ignored tests that were no longer
necessary.