DEV-972 Replace transfer API-based archive
There were too many administrative hurdles to using the previous
implementation so replace it with rclone. The same remote is used
(but a different bucket) so there are minimal changes and the new
code is actually very simple.

One of the reasons for using the transfer API to start with was
reluctance to trust everything to rclone, which is what we're now
doing (though the transfer API was being trusted to do everything
in that case). I'm pushing this functioning version for some feedback
as I'm not sure whether we might want to do more explicit validation
after the files are in the archive location.

Also when we were testing the transfer API implementation the config
of the development storage buckets was amended to make it more
production-like. Due to this the smoke test began to have the same
error we had seen in staging and I elected to update the code so
that the manifest is copied directly to S3 rather than jumping through
GCP first, to avoid the problem.