DEV-1026 Make first cut at CRAM batch converter
This is a pretty simple wrapper around our existing VM spawning code
that makes it easy to spool up multiple VMs in parallel. It is
designed to take a file containing a list of input files that will be
operated upon (and passed to each VM in turn). For the moment there is
no special handling of the results; the only requirement is that the
operation returns a `PipelineStatus` object.

There are some really rough edges here, just wanting to get something
pushed and some eyes across it so I can get some feedback before going
down any rabbit holes.