DEV-1026 Move batch code to new module
We decided that there is enough here to make a new module. For now the code
that is used by both `cluster` and `batch` can stay in `cluster`, but that
is prone to change at some point.

This commit also adds operation selection on the command line via some
reflection and a simple command registry. The action that is being
requested will need to be the first token on the command line.

Changes have been made to clean up the output, and now there is a single
bucket for all of the output; also any logs and metadata is namespaced
according to the instance id of the VM on which it originated, within that
single output bucket. An adjustment was made to the `OutputBucket` class to
fix a problem with directory creation that meant the first VM's output and
metadata would end up in a different directory from that of all subsequent-
finishing VMs.

Finally an implementation of the batch Flagstat generation command which
is part of DEV-1025 snuck into this commit as I was working on the
framework at the same time as I was writing it.