DEV-1032 Fix issues with pre-emption during vm start and error prop.
Checks for the "UNSUPPORTED_OPERATION" error and when encountered, logs
a warning and goes to next zone. Basically the same logic as zone
resources exhausted.

Throw a runtime exception whenever a file is missing compiling the final
results.

The FullPipeline was not checking single sample state past the alignment
We know grab the final state of both single sample pipelines before
reporting the final status.