DEV-972 Make gsutil output more on failure
Also remove the threading that parallelises the pushing of the output to
both the SBP cloud and the GCS archive as errors in the archive portion
will not prevent the SBP push.

Also remove the verbose output from the GSUtil class, as it is
dangerous (credentials may be exposed by it) and doesn't really give
us much benefit.