DEV-1045 Upgrade to gridss 2.7.2 and use driver script
The upgrade is relatively straightforward, just a version bump.

Moving to the driver script is a huge simplification. We can remove all
our own implementation of the gridss pipeline and just give it inputs
and consume outputs into the somatic filtering and annotation. This
commit removes much of the custom gridss code.

Also of note we include samtools as the driver script prefers this to
sambamba for greater compatibility across CPU types.