DEV-1026 Require output bucket to exist
What we had was fine for the development case in which the source and
output buckets would be in the same project. However after going through
the process of getting something working in a non-development environment
it becomes clear that it is preferred to have the output in a
less-privileged project and the input from a more-restricted project. For
instance the patient data may come from the production datastore but the
compute and output should be in a sandbox. For that use-case it makes sense
to allow the production service account to write to the sandbox
environment, and the only real way to set up the permissions for that
properly is manually before the run.