DEV-1060 First stab at versioning images
Changes the image name from "diskimager" to "pipeline5" and uses a
version instead of a "type" (was always standard anyway).

The code references the image via a hardcoded version for now, as
dynamic version resolution would cause issues with local-SNAPSHOT. This
can hopefully be improved when we tie the image and the build.