DEV-1025 Automatic comparison of flagstats
The original version of the operation just created a flagstat file from
the input BAM, but this version also copies down the flagstat from the
original run and compares it to the newly-generated one, failing the
operation if they do not match exactly.

Also remove the "storage key" argument as we have evolved the setup on
the GCP side to no longer require it. And add a README for the batch
framework.