DEV-1026 Provide Sambamba-based converter
Sambamba doesn't really help performance and in practice we've seen some
differences in the flagstats for the BAMs it produces so we probably
won't be using it for the batch conversion. Nonetheless it may be useful
to have this operation in the near term.

Also adjust VM parameters for samtools conversion based on empirical
evidence.