DEV-965 Use users account if no private key specified
Makes the private key an optional argument, and if not present will use
the users own account to interact with the GCP apis. Note the service
account email is still required.

Not the cleanest use of Optional but as the credentials throw an
exception, this looks better than using map or ifpresent.