DEV-1087 Update README.md and add scripts
The scripts are just thin wrappers around the GCP commands to export the
images and then import them into another project, which is the new way
of doing promotions since we moved the resources into the image.

The README.md had a significant rewrite as it had been drifting for some
time. Also a new section was added to help clarify the development
process.