DEV-1092 Allow VM to specify disk requirements
Also decouple the disk size from the type so that whether local SSDs
or persistent storage is chosen an appropriate size will be selected
based upon the requirements of the job definition.

Also slightly amend the Bam2Fastq job to take advantage of the changes
and also hike its memory as some jobs were failing due to GC overhead
exceeding the setpoint.