DEV-1094 Upload dispatcher log on completion
Also improve the logging to include the versions as we do with the
pipeline. Further the mapping between the VM label ("number" in the
batch) and the input is now pushed to the runtime bucket at startup
so the operator can get an idea of what is going on while the run is
still happening.

The namespacing for the output has been refined so that all of the
output for a particular VM instance goes into its own directory rather
than one big directory as that does not work very well for involved
outputs.