DEV-854 Refactoring of initial implementation
Separates the aggregating of results, copying of output, and the interaction with SBP API from
the main bcl2fastq class. Most of the logic is actually in the combining of the sample sheet,
resulting fastq files and Stats.json. By doing this up front, the post-processing becomes a lot
more straightforward and easy to test.

Simplifies the QC a lot by removing the interface and separation from SBP API code. Makes sense
to keep them together as they are tightly coupled (and quite inconsistent with how each check
is propagated).