DEV-1106 Always use a second PD SSD when local SSDs disabled
This solves the issue of the 2TB limit of the boot disk. The boot disk will
always be 100GB (although perhaps that can be reduced?) and workind disk size
the size of the new disk.

The disk is mounted in the startup script with the PersistentStorageStrategy.
this mounts /dev/sdb to /data.