DEV-1142 Include undetermined yield in Q30 calculations
A bit of a refactor to make things more consistent. We create an object in our Conversion datamodel to hold
the undetermined stats. This is populated from read metrics now deserialized in the UndeterminedStats

All stats in Conversion are now derived, to ensure the numbers are always consistent across the datamodel.