DEV-1122 Don't copy new BAMs after CRAM finishes
The CRAM conversion generates a BAM from the new CRAM and then
compares that to the original. These are not generally useful
to copy up to the cloud, but we do want to keep them around if
there is anything out of ordinary. Move them to another folder
so they aren't automatically copied up on job completion.