DEV-1115 Store important input and output files for forensics
Will copy a list of blobs both from the runtime results path and the input path into a special bucket
for later analysis of issues. See JIRA for the complete list.

The bucket is configured as an argument. Each conversion gets its sub-folder in the same bucket.

The log of the actual Bcl2Fastq is stored as well, so we add a log4j config here and copy the output of
the FileAppender into a blob called conversion.log.