Improve disk imager speed
While we do not do imaging often it was getting to be 90+ minutes which
is too long. By adding cores to the imaging VM, ensuring it is using
SSDs and updating the `gsutil` command to make use of multiple cores the
elapsed time is down to ~40 minutes.

Also increase the size of the disk image as we were getting close to
running out.