DEV-1245 Create a JSON sample source
Create a new class JsonSampleSource which parses a JSON representation of the sample/lane/fastq model,
passing the GCS locations along as paths within the lanes. This combined with the GS upload allows us
to put fastq files in arbitrary locations.

Eliminated the file system upload as its no longer used (local runs should use this new way, or the acc
api).

Also eliminated the patient module and some interfaces which didn't really add any value. Included in this
commit as it made the addition of the JsonSampleSource more straightforward.