DEV-1279 Copy all single-pipeline files on rerun
We encountered a failure in production in which a failure in the CRAM
stage resulted in a rerun of the pipeline, but after the rerun succeeded
the output directory did not contain the files from the sample that had
failed the first time (none of them; the subdirectory for the sample
itself wasn't even present).

Looks like the reason for this was that the flag file which tells the
pipeline that the staging directory has been populated was left-over
from the first run, which meant that upon re-run the pipeline didn't
wait for the files to be copied from the staging bucket into the final
archive bucket.

Also improve logging output for post-applied operations to make
debugging this problem easier.