DEV-1287 Disable sage caller when shallow run
Currently the output is not used in the shallow run and it doesn't provide anything interesting to use
later. Disable the caller when the shallow flag is true.

Also added a more general test in StageTest for shallow mode. This way the subclasses can declare whether
they are enabled for shallow with a simple boolean override.