DEV-1289 Validate the fastq naming if we're running in production
After some discussion decided that when we run against production samples its best that we
ensure the fastq name matches our expectations. This way we'll never quietly push jumbled
RGs into crams.

Uses a regex to validate the format, but continues to just remove R1 and R2 from the string
as it has the same result as recontructing the name, with less code.