DEV-1342 Add turquoise event publishing
Adds publishing of an event on pipeline start and pipeline completion with the following subjects:
- sample
- set
- type (somatic, shallow, single)
- barcodes
- run id if present
- status for completion

For now made the choice to completely decouple from turquoise at compile time, and duplicate the
classes required to publish the events. Will probably revisit this later and perhaps create a simple
turquoise client.

The publishing is hid behind a feature toggle which is for disabled.