Update imager packages
Sometime in the last little while we'd lost the ability to build the
image using the existing packages due to upstream changes.

Switch to a later version of R that is still supported and also switch
to the adoptopenjdk Java 8 version as Debian had removed OpenJDK8,
apparently for security reasons. This last change required adding a new
third-party repository and more bootstrapping software.