Rollback to Debian 9
Previously this was the default and I didn't realise that Google had
promoted Debian 10 to be the default. Now explicitly use Debian 10.

I've been able to confirm that the original packages continue to work
when the OS is rolled back so revert the changes to the packages made
earlier too.