DEV-1443 Create a profile for platinum Arguments
The profile is called PUBLIC, which in this context means it is a profile which can
be shared publicly. Of course the intention it will always be invoked via platinum,
but maybe that won't always be the case.

First clean up all unused arguments and classes which still reference them. Also rename
private network to just network (private network isn't really a thing) and patient
report bucket to output bucket (no reports are in that bucket).

CMEK becomes an optional again, and the strict checking is disabled.

The public image is just enabled by a flag, which is actually not exposed to the CLI
and can only be enabled with the PUBLIC profile. This way we can never have public
mode running with a private image