DEV-1476 First cut at resource overrides
This commit updates our imaging scripts to implement a two-stage model.
Stage 1 images are suitable for distribution to third parties and do not
contain any resources. The Stage 2 image would be the same as what we
currently use in production.

Also a new ResourceFiles implementation has been introduced that uses
a bucket to provide the resources to the running pipeline rather than
getting them from the image. A JSON file is provided by the user to
describe the resource locations within the bucket. The performance of
this solution is quite acceptable with the current size of the resources
and VMs that we are using in the pipeline. Potentially it could have a
relatively larger impact on runtime if extremely-small VMs are being
used but it does have the nice benefit of not requiring the user to do
any imaging.