DEV-1478 Run from persisted alignment data (WIP)
First step in re-running from any stage is to re-run from alignment. This was kind of supported
in the past with "AlignmentOutputStorage" but this approach prescribed a fix location in the
runtime bucket for the bams and could not support crams.

The new way hopefully is a nice template for the rest, which introduces a new interface for the
aligner, with a new persistent implementation. The new PersistedAlignment will look for the
existing alignment in the output bucket and use these files in place as the input for subsequent
stages.

This involved various refactoring to support. First, superficially needed to change the name of the
"VmAligner" to BwaAligner to more accurately describe its specificity.

The SomaticPipeline is changed to take an AlignmentPair as input, rather than looking it up via the
storage. This allows us to pass either the "live" alignment or a stored one without any awareness
in the pipeline.

Cram is also supported, but still need to test. Not sure if all tools can work directly with a cram
without additional arguments.