DEV-1478 Run from various starting points
Adds the ability to specify a starting point to the pipeline. The starting point can be
either alignment, calling or purple complete. Each starting point builds off the last,
so purple complete is the union of all 3.

For each stage which can use persistent data, there is a PersistedXxx implementation.
These classes contain the knowledge of the location of the output files for outputs
created with this same version of the pipeline.

It works by using the exists "skipping" mechanism, but enriching the skipped output
with the actual location in google storage.

The remaining "OutputStorage" classes which used to poll for metrics and germline data
from the single sample pipelines are replaced with queues to avoid any duplication with
the persisted implementations.