DEV-1488 Use JSON sample names
We now use the sample names in the JSON file rather than deriving the
names from the set name. This improves usability in Turquoise, in which
we were getting failures if the set name and sample names didn't
"agree".

Also make the pipeline fail with a non-zero exit code upon argument
parsing failure.