DEV-1498 Use persisted data for single sample pipeline stages
Adds the concept of a persisted dataset which can be used as a directory of
files, looked up by datatype, run and sample. This first version only supports
the single sample stages of alignent, bam metrics and germline calling.

The stages generally follow the pattern of falling back to the conventional
location of the files if they can't be found (persistentDataset.find().orElse())

Also adds fields to the run metadata for the id of the run, the bucket, and
the set name to make this lookup process a little easier and consistent
throughout.

Refactored a bit to consolidate file types and data types, in particular the
addition of the index extensions which was spread in a few places. Also removed
the sorted from all bams as it no longer serves purpose, and no sense carrying
it forward to the new structure.