DEV-1498 User persisted data for somatic stages
Refactors previous change to also support looking up somatic data from persistent storage.

By adding the barcode to the RunMetadata we can actually avoid having a different implementations
and look up the files using the tumor barcode.

As many of the HMF tools just need a directory, we add a second method to the dataset interface.
Maybe later we refactor the model in the API to better support this, so nice to hide it behind
and interface.

Add a simple check in the cleanup to ensure we don't fail deleting runtime buckets which do not
exist.