DEV-1547 Run from bam files specified in JSON and allow overriding of ref genome
Bam support is added via the JSON datamodel, with an optional path. If the path is there, lanes
are ignored and the BwaAligner just returns the location of the bam to subsequent stages.

Overriding the ref genome is done with a new argument (this will be a bit more convenient in
platinum, as including it the JSON will mean a lot of duplication). If the argument is present,
the resource files are decorated and a special command is added to the bash script to download
the genome to the normal resources directory.