DEV-1555 Print failure logs at the end of pipeline
Adds a field to the stage outputs to cache a reference to the log blob in case of failure.

Little awkward as we need to explicitly add the log everywhere, but can't find another solution
with the current state of the StageOutputs. The lanes in the BwaAligner also make it a bit
tricky to make a really general solution.