DEV-1593 Add biopsy, new metadata api, and new dataset
The biopsy name can now be passed as an argument, which puts the pipeline in a "rerun" mode. This is done
with a special metadata api and dataset api.

The metadata api just uses the biopsy to resolve a tumor and ref sample (filters rna). Currently multi-dna-set
biopsies are not supported. Also the completion method is not yet implemented.

The dataset api is really simple now, just a map of the response from the rest api, as it handles resolution
of the right versions, sets and samples.