DEV-1645 Automate tool upgrade, image creation and docker image
Simple script to copy the new jar into the common-tools bucket and start the image creation process. This now
is broken up into two sets of commands: base and tools. Passing the -t flag will create an image from the
previous build in the family, only overlaying the tools dir.

Otherwise the upgrade script does a maven build to create a docker container and push it to GCR (rather than
docker hub).