DEV-1707 Increase disk for tumor CRAM conversions
There was a failure in production because of disk space limitations on
an unusually-large tumor sample.

Also since the initial writing of the pipeline I believe GCP relaxed
their requirement that SSDs be provisioned in pairs, so I've also
removed the code that bumped that up if the calculated value was odd.
As a consequence of this, I have also upped the default for disk space
from 900MB to 1200MB as 900MB as very few jobs specify a non-default
number for the disk space. Previously, 900MB would be calculated to 3
disks but rounded up to 4 and I don't know what the jobs that don't
specify an explicit value, actually need.