DEV-1775 Distinguish between file/dir datatypes
Rather than create the datatype add operation with an empty filename to
indicate a directory, make two static methods on the class and make it
harder to call one when you mean the other.

Also rework the existing Cobalt and Amber datatype applications to have
more-suitable names and drop the '_output' suffix from Linx and
Bachelor.