DEV-1786 All add datatypes use a path to an actual file
Extracted the path creation logic into a simple path class to remove some duplication and make it
a little clearer what AddDatatype accepts.

Otherwise linx we use the first of the tsvs in alphabetical order, and bachelor we use the germline
variant tsv as that is where the "main" data resides.