DEV-1798 Support HG38 Amber PON after all
The resources now contain a VCF so that the pipeline run will finish
if done under HG38. That VCF is currently a husk with no actual data in
it so the subsequent patient DB load will not actually put anything into
the database, however this will be resolved when a proper VCF for 38 is
available.