DEV-1833 Publish staging events when pipeline complete in rerun mode
Adds a new implementation of the complete method which publishes a PipelineStaged event when the pipeline has
not failed. This involves iterating through the output and passing each path to the event,along with its
optional data type. The datatypes are now linked in the event subscriber.

Due to this, it was practical to remove the interface which AddDatatype implemented (no other implementations).
Eventually will be able to simplify it further once live pipelines are archived this way.

Jackson version needed bumping to be compatible with the new event api.