DEV-1931 Split public and private images
I've reorganised the imaging scripts to have really only two entry
points: one to create the public image, and the other to create a
private image. The private image will contain our internal-only
resources.

The private image will always be created out of a public image. The new
script will overlay a public image with our private resources, then cut
a new image with the same name. However it uses separate projects on the
GCP side to ensure that only some users are able to save the new images,
and similarly to restrict access to the private resources.

Also this includes some cleanup to the scripts and README.