DEV-1931 Most SAGE resources moved to repo
Previously all of the SAGE resources were hosted in the bucket because
the overall size is quite large. However the hotspots need to be updated
often and only the germline PON is large.

So now the imager cherry-picks the files from the subdirectories that
are still in the bucket. This will do for now.