DEV-1941 Rework handling of repo/bucket resources
The base commands have been amended to just copy everything from the
bucket after the git repo is fetched from. The `-n` flag has been
provided to `gsutil` in order to prevent it overwriting anything that
was already cloned from the repository. In that situation a warning
is issued by `gsutil` which should be good enough.

Further a few more commands were added to un-gzip some of the new
resources which were stored un-compressed in the bucket but compress
well and have been compressed in the repository. It is worth keeping
the repository as small as possible for usability and maintenance
reasons so this seems a good trade-off. Also many of the text-based
resources are already stored compressed so it's not a big departure.