DEV-1949 Fix smoke test expectations
The expected files needed to be updated to include the PEACH output files, as
well as the Linx driver catalog as 1.15 now outputs this even if it has
no content. This change in Linx 1.15 now means that PROTECT also runs properly
so its output files were added too.

Also fix the ref/tumor names passed to PEACH, it was getting the barcodes,
which happen to be the same as the sample names in our test data.