DEV-1932 Update Cuppa to use new Python framework
Cuppa charts are now generated using the improved mechanism for Python
environments.

Also remove the registration of the chart as a datatype, turns out that
wasn't needed.

Tests have been updated also, however one continues to fail as it is
testing the inputs, which still include the viral insertions from Linx.
This was removed from the Linx output along with the update to PROTECT
1.3 but results in a "soft failure" (error logging but still an exit
with 0) from Cuppa so I am leaving it in for now so this doesn't get
forgotten.