DEV-1932 Update Cuppa to use Linx output dir
Linx has been modified to include its entire directory in its output object
and Cuppa now downloads that entire directory rather than picking out files
one at a time. As Cuppa uses several files from Purple too (such as the QC
file) that are also expected to be in the input directory but which are not
explicitly mentioned on the Cuppa command line, I have switched to using the
Purple outputdir for those as well.