DEV-1970 Pull cuppa-chart out of the Cuppa JAR
Rather than a standalone tarball the cuppa-chart script is now packaged as a
resource in the Cuppa JAR. So rather than unpacking and installing from that
tarball at imaging time we will unpack just the resources from the JAR and
then proceed in the same way.

Also remove the separate CUPPA_CHART version from the pipeline.

While I was updating the imaging commands I eliminated the naive "pipe to bash"
installation mechanism for `pyenv`, which while convenient is not overly
secure. There is now a tarball in the tools bucket which can be updated if
needed and instructions for how to do so are in the imaging commands.