DEV-2082 Add datatype to the persisted output of bam metrics
This will be picked up by the StagedOutputPublisher even though the stage has not actually been run. Then the archiver
will receive the datatype and barcode and add the proper file api links.

As bam metrics are the only non-cram files grouped into the secondary staged event, we only pull this trick in one place.
Longer term it would be good to factor the code such that persisted results always honor the datatype of the normal
output generation.

Test plan:
- Run research pipeline in pilot
- Ensure that the file is persisted with api objects
- Run DB loader and check that MetricsLoader is run.