DEV-2096 Cache key for R repo locally
We are getting intermittent failures on our imaging machine because the
key server is not always responding. Not sure why this is but I have
made a copy of the key from a successful run and we can just copy it
at startup.

I've also changed the copying of the `mk_python_venv` script so that it
no longer goes via a bucket, that seemed a bit untidy and the new way
that is used with the key is nicer.