DEV-2086 Add datatypes for patient-reporter
As of this writing these are handled by a temporary class in the wrapper
code for `patient-reporter` which registers the missing files in the API
but we'd like to get rid of that class when we've got everything
registering properly from here.

Also rename some of the existing datatypes to remove their "TSV" suffix,
which isn't really consistent with the others.