DEV-2197 R dependencies for Debian 10
This file is not yet used but when we are ready to move to Debian 10
from the already-deprecated Debian 9 for the pipeline we'll need to
rebuild the R libraries.

R libraries are coupled to the environment they were compiled in and our
current production version uses cran35 while the new one uses cran40.