DEV-2217 Fix up problems with imager
There were a few problems here: (1) Another PR for PEACH was created
uner this same branch name to update one of the dependencies, as the
existing version had stopped building properly.

Potentially frustrating things with the dependencies was the fact that
the `pip` command was not being properly propagated into the PyEnv
environments unless the copy that we extract from our tarball into
`/root/.pyenv` was first updated. The symptom was that the `pip` command
would be namespaced under Python version 3.5 instead of 3.6 in the
virtual environment, which if nothing else was confusing.

I've added some cleanup commands in the `tools.cmds` to delete existing
PyEnv and tools as problems are caused by having the Python environments
already in place when running with the `-t` command.

Also bump the versions to avoid confusion with the existing 5.23 images.